adds ids to inputs
